firstboot: trigger LUKS2 reencryption properly
Since there are two key slots in the LUKS partition, one for the
recovery password and the other for the random key, it needs to specify
a key slot number with "--key-slot" when reencrypting the partition.
To trigger reencryption with "--key-slot", this commit checks whether the
user chooses to use a password or not. If yes, drop the random key slot
before reencryption. Otherwise, drop the password key slot.

Signed-off-by: Gary Lin <glin@suse.com>