Make '--passfile' compatible with binary files
Originally fde_request_recovery_password copied the content of the file
from '--passfile' to the global variable, result_password. Since the
user may provide a binary file as the password file, direct copying to
the variable may not work. To make the option compatible with binary
files, all the callers of fde_request_recovery_password switch to
fde_request_recovery_passfile to output the password from '--password'
or '--passfile' to the given file, so that the underlying commands, such
as cryptsetup, can read the binary key.

Signed-off-by: Gary Lin <glin@suse.com>