firstboot: Update grub config after removing the key file
When KIWI creates the image, the LUKS key file is created for the
firstboot, and the FDE firstboot scripts removes the key file later.
However, since the key file exists during the generation of the image,
grub2-mkconfig omits the creation of the 'crypttab_entry' commands in
/boot/grub2/grub.cfg. To create the 'crypttab_entry' commands correctly,
we have to update /boot/grub2/grub.cfg right after removing the key
file.

Signed-off-by: Gary Lin <glin@suse.com>