Add fde-tpm-helper for bootloader RPMs
To update the authorized policy automatically after a bootloader
upgrade, the fde-tpm-helper script is created to detect if the system
needs to update the signature in the sealed key. It requires the
bootloader RPM spec to add %fde_tpm_update_post to %post and
%fde_tpm_update_posttrans to %posttrans.

An additional sysconfig variable, FDE_TPM_AUTO_UPDATE, is introduced to
allow the system admin to decide whether to enable auto-update or not.

Signed-off-by: Gary Lin <glin@suse.com>